+++ app/(tabs)/crew/_layout.tsx
@@ -2,7 +2,7 @@ import { Stack } from "expo-router";
 
 export default function CrewLayout() {
   return (
-    <Stack>
+    <Stack initialRouteName="home/index">
       <Stack.Screen name="home/index" options={{ headerShown: false }} />
       <Stack.Screen name="detail/[id]" options={{ headerShown: false }} />
     </Stack>
